// admin/src/api/generated/models/DeploymentEventResponse.ts
export interface DeploymentEventResponse {
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    createdAt: string;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    eventName: string;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    failureStage?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    message?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    module?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    moduleName?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    stage?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    status?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    step?: string | null;
    /**
     *
     * @type {string}
     * @memberof DeploymentEventResponse
     */
    stepId?: string | null;
}